refactor(imports): update import paths for SectionContainer and SectionHeader components
